warning as error (#27)
Squashed commit of the following:

commit b32b6cd38adc22603f1a6ed98b4e22a072bf550c
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 12:14:01 2016 +0800

    fix unsigned comparison

commit 4b92d01c2a9f1816bdf76d69c9aa5888322cd71b
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:41:21 2016 +0800

    use unused variable

commit 31ebe40a7554c7c1cacfeca5b8d9c492b1ad5101
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:37:31 2016 +0800

    don't used uninitialized value

commit fbfde4350965beaed451aded73c563f4e3d8bb2c
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:35:53 2016 +0800

    use unused variable

commit 05e652bb3b10d6c36e5271b06f156b753dcc25b2
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:34:33 2016 +0800

    fix unsigned comparison

commit d643cb3312f603c7413c4557714cf7ae93a35fe4
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:28:54 2016 +0800

    add paren

commit 7d5554d304c0625fb732734961c7222e13780354
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:27:27 2016 +0800

    fix comparison between signed and unsigned

commit 6a065b8bfa458ed6aa55fcb49e5ee5a904ef65da
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:20:33 2016 +0800

    use set-but-not-used value

commit e1977f190dd02da1f7739ac7bf62da0b46ba2851
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:07:54 2016 +0800

    don't throw in constructor, will always terminate

commit df35784be10dff9bb9af4af289393a5dc91d43c4
Author: ahxxm <i@ahxxm.com>
Date:   Mon May 16 09:05:40 2016 +0800

    fix comparison between signed and unsigned

commit fd97a3d334434dd9ab7fd879e3755e1c2b00c433
Author: ahxxm <i@ahxxm.com>
Date:   Sun May 15 12:51:34 2016 +0800

    fix comment about dangling pointer

commit 307ac7d121168f509bd3a09ddef38ecd7c774021
Author: ahxxm <i@ahxxm.com>
Date:   Sun May 15 12:47:27 2016 +0800

    try warning as error

+++ cppp/src/9-sequential-container/27-28-29.cpp
@@ -62,7 +62,7 @@ TEST(ForwardListTest, SomeTest) {
 TEST(VectorResizeTest, SomeOtherTest) {
   // 29:
   std::vector<int> va(25, 10);
-  EXPECT_EQ(va.size(), 25);
+  EXPECT_EQ(static_cast<int>(va.size()), 25);
 
   va.resize(100);
   EXPECT_EQ(va[99], 0);
